Moved scoutmanager to the Modules folder
