+++ Python/Modules/scoutmanager.py
